Re-organising of Framework class files
Moved code specific commands into Codes/
Moved Framework lattice and object functions into Framework_objects.py
Moved Framework element functions into Framework_elements.py
Moved online model plotting functions to the online_model repository
