+++ SimulationFramework/Examples/FEBE/transverse_best_changes_upto_S07.yaml
@@ -1,39 +1,39 @@
 CLA-S02-MAG-QUAD-01:
-  k1l: -0.18894543101287667
+  k1l: 0.01286637356941942
 CLA-S02-MAG-QUAD-02:
-  k1l: 0.029889008599559938
+  k1l: -0.05514991629939368
 CLA-S02-MAG-QUAD-03:
-  k1l: 0.11156353738891792
+  k1l: 0.05104272046233741
 CLA-S02-MAG-QUAD-04:
-  k1l: 0.16820212433408527
+  k1l: -0.005985202906625358
 CLA-S02-MAG-QUAD-05:
-  k1l: 0.1397671398738332
+  k1l: 0.04359924522567179
 CLA-S03-MAG-QUAD-01:
-  k1l: -0.22025904398502594
+  k1l: -0.08568957392797992
 CLA-S04-MAG-QUAD-01:
-  k1l: 0.15009580132438638
+  k1l: -0.09535925963950304
 CLA-S04-MAG-QUAD-02:
-  k1l: 0.07791883767232
+  k1l: 0.06349939430222948
 CLA-S04-MAG-QUAD-03:
-  k1l: -9.179166136759662e-05
+  k1l: 0.23837331036691262
 CLA-S04-MAG-QUAD-04:
-  k1l: 0.08558653580731868
+  k1l: -0.03497565552521683
 CLA-S04-MAG-QUAD-05:
-  k1l: 0.11063165876713685
+  k1l: -0.17257624456138557
 CLA-S04-MAG-QUAD-06:
-  k1l: -0.09479599523297264
+  k1l: 0.0181310974153476
 CLA-S04-MAG-QUAD-07:
-  k1l: 0.11057345939252969
+  k1l: -0.13185879324014435
 CLA-S04-MAG-QUAD-08:
-  k1l: 0.04801546796966104
+  k1l: 0.02998695552911902
 CLA-S05-MAG-QUAD-01:
-  k1l: -0.07260732671120324
+  k1l: 0.25496069506762714
 CLA-S05-MAG-QUAD-02:
-  k1l: 0.296423602639331
+  k1l: 0.07290117271274413
 CLA-S06-MAG-QUAD-01:
-  k1l: 0.11373580579574802
+  k1l: -0.013660882162709463
 CLA-S06-MAG-QUAD-02:
-  k1l: -0.1471430939449066
+  k1l: 0.
 CLA-S07-MAG-QUAD-01:
   k1l: 0.362686
 CLA-S07-MAG-QUAD-02:
